bpl_cfg: fix cfg_get_link_metrics_request_interval
Read LinkMetricsRequestIntervalSec as uint32_t instead of int64_t to match the
datamodel parameter type.

Closes: PPM-3970
References: FEAT-319

+++ framework/platform/bpl/cfg/dm/bpl_cfg_controller.cpp
@@ -232,7 +232,7 @@ bool cfg_set_link_metrics_request_interval(std::chrono::seconds &link_metrics_re
 
 bool cfg_get_link_metrics_request_interval(std::chrono::seconds &link_metrics_request_interval_sec)
 {
-    int64_t interval_sec = 0;
+    uint32_t interval_sec = 0;
     if (!read_controller_config_param("LinkMetricsRequestIntervalSec", interval_sec)) {
         return false;
     }
